Misc PR adjustments; add + integrate dynamic language translations; cleanup.
