+++ context/AppContextProvider.tsx
@@ -11,7 +11,7 @@ import {
 } from 'react';
 import axiosClient, { setAuthToken } from '@/config/client';
 import { onIncompletePaymentFound } from '@/config/payment';
-import { MembershipClassType } from '@/constants/membershipClassType';
+import { MembershipClassType } from '@/constants/mock';
 import { AuthResult } from '@/constants/pi';
 import { IUser } from '@/constants/types';
 import logger from '../logger.config.mjs';
