+++ messages/de.json
@@ -51,7 +51,13 @@
       "CURRENT_MEMBERSHIP_END_DATE_NO_ACTIVE_MEMBERSHIP": "Keine aktive Mitgliedschaft",
       "MAPPI_ALLOWANCE_REMAINING_LABEL": "Verbleibendes Mappi-Guthaben",
       "PICK_MEMBERSHIP_MAPPI_TO_BUY_LABEL": "Mitgliedschaft oder Mappi zum Kauf auswählen",
+      "PICK_MEMBERSHIP_DURATION_IN_WEEKS_LABEL": "Mitgliedschaft ({duration} Wochen)",
       "PICK_BUY_METHOD_LABEL": "Kaufmethode auswählen",
+      "PURCHASE_OPTIONS": {
+        "PAY_WITH_PI": "Mit π bezahlen",
+        "WATCH_ADS": "Werbung ansehen (kostenlos)",
+        "USE_VOUCHER": "Gutscheincode verwenden (kostenlos)"
+      },
       "ENTER_VOUCHER_CODE_PLACEHOLDER": "Gutscheincode eingeben",
       "VALIDATION": {
         "SUCCESSFUL_MEMBERSHIP_ACTIVATION_MESSAGE": "Mitgliedschaft erfolgreich aktiviert",
