+++ messages/en.json
@@ -51,7 +51,13 @@
       "CURRENT_MEMBERSHIP_END_DATE_NO_ACTIVE_MEMBERSHIP": "No active membership",
       "MAPPI_ALLOWANCE_REMAINING_LABEL": "Mappi allowance remaining",
       "PICK_MEMBERSHIP_MAPPI_TO_BUY_LABEL": "Pick membership or Mappi to buy",
+      "PICK_MEMBERSHIP_DURATION_IN_WEEKS_LABEL": "membership ({duration} weeks)",
       "PICK_BUY_METHOD_LABEL": "Pick buy method",
+      "PURCHASE_OPTIONS": {
+        "PAY_WITH_PI": "Pay with π",
+        "WATCH_ADS": "Watch ads (free)",
+        "USE_VOUCHER": "Use a voucher code (free)"
+      },
       "ENTER_VOUCHER_CODE_PLACEHOLDER": "Enter voucher code",
       "VALIDATION": {
         "SUCCESSFUL_MEMBERSHIP_ACTIVATION_MESSAGE": "Membership activated successfully",
