+++ messages/ewe-BJ.json
@@ -51,7 +51,13 @@
       "CURRENT_MEMBERSHIP_END_DATE_NO_ACTIVE_MEMBERSHIP": "Míele nyaɖe me le ɖoɖo la ta",
       "MAPPI_ALLOWANCE_REMAINING_LABEL": "Mappi si gblɔ na wo kple kple le",
       "PICK_MEMBERSHIP_MAPPI_TO_BUY_LABEL": "Tia nudɔwɔlaƒe alo Mappi be nàɖɔ",
+      "PICK_MEMBERSHIP_DURATION_IN_WEEKS_LABEL": "dzidzɔla ({duration} kɔsiɖa)",
       "PICK_BUY_METHOD_LABEL": "Tia ɖɔɖɔ nu vovotɔ",
+      "PURCHASE_OPTIONS": {
+        "PAY_WITH_PI": "Tɔ π",
+        "WATCH_ADS": "Kɔ ads (free)",
+        "USE_VOUCHER": "Tɔ voucher code (free)"
+      },
       "ENTER_VOUCHER_CODE_PLACEHOLDER": "Ɖe voucher kɔda me",
       "VALIDATION": {
         "SUCCESSFUL_MEMBERSHIP_ACTIVATION_MESSAGE": "Membɛship la wótsɔ ɖe eme gɔ̃ hã",
