+++ messages/hi.json
@@ -51,7 +51,13 @@
       "CURRENT_MEMBERSHIP_END_DATE_NO_ACTIVE_MEMBERSHIP": "कोई सक्रिय सदस्यता नहीं",
       "MAPPI_ALLOWANCE_REMAINING_LABEL": "शेष Mappi भत्ता",
       "PICK_MEMBERSHIP_MAPPI_TO_BUY_LABEL": "सदस्यता या Mappi खरीदें चुनें",
+      "PICK_MEMBERSHIP_DURATION_IN_WEEKS_LABEL": "सदस्यता ({duration} सप्ताह)",
       "PICK_BUY_METHOD_LABEL": "खरीदने की विधि चुनें",
+      "PURCHASE_OPTIONS": {
+        "PAY_WITH_PI": "π के साथ भुगतान करें",
+        "WATCH_ADS": "विज्ञापन देखें (फ्री)",
+        "USE_VOUCHER": "वाउचर कोड का उपयोग करें (फ्री)"
+      },
       "ENTER_VOUCHER_CODE_PLACEHOLDER": "वाउचर कोड दर्ज करें",
       "VALIDATION": {
         "SUCCESSFUL_MEMBERSHIP_ACTIVATION_MESSAGE": "सदस्यता सफलतापूर्वक सक्रिय कर दी गई है",
